+++ packages/contracts/contracts/test/harnesses/TestBalanceManager.sol
@@ -25,12 +25,13 @@ contract TestBalanceManager is IHandler, BalanceManager {
     // Stub to make testing between Teller<>BalanceManager easier
     function handleDeposit(
         Deposit calldata deposit
-    ) external override onlyTeller {
-        _handleRefundNote(
-            deposit.encodedAsset,
-            deposit.depositAddr,
-            deposit.value
-        );
+    ) external override onlyTeller returns (uint128 merkleIndex) {
+        return
+            _handleRefundNote(
+                deposit.encodedAsset,
+                deposit.depositAddr,
+                deposit.value
+            );
     }
 
     function handleOperation(