+++ packages/contracts/contracts/test/invariant/actors/HandlerHandler.sol
@@ -10,7 +10,6 @@ import {console} from "forge-std/console.sol";
 import {TokenSwapper, SwapRequest} from "../../utils/TokenSwapper.sol";
 import {TreeTest, TreeTestLib} from "../../utils/TreeTest.sol";
 import "../../utils/NocturneUtils.sol";
-import "../../utils/ForgeUtils.sol";
 import {Teller} from "../../../Teller.sol";
 import {Handler} from "../../../Handler.sol";
 import {ParseUtils} from "../../utils/ParseUtils.sol";