+++ packages/sdk/test/utils.ts
@@ -12,6 +12,7 @@ import {
   IncludedNoteWithNullifier,
   WithTotalEntityIndex,
   range,
+  IncludedNote,
 } from "../src";
 import { Handler, Handler__factory } from "@nocturne-xyz/contracts";
 import randomBytes from "randombytes";
@@ -99,13 +100,15 @@ export async function setup(
   const nocturneDB = new NocturneDB(kv);
   const merkleProver = new SparseMerkleProver(kv);
 
-  const notes = zip(noteAmounts, assets).map(([amount, asset], i) => ({
-    owner: signer.generateRandomStealthAddress(),
-    nonce: BigInt(i),
-    asset: asset,
-    value: amount,
-    merkleIndex: i,
-  }));
+  const notes: IncludedNote[] = zip(noteAmounts, assets).map(
+    ([amount, asset], i) => ({
+      owner: signer.generateRandomStealthAddress(),
+      nonce: BigInt(i),
+      asset: asset,
+      value: amount,
+      merkleIndex: i,
+    })
+  );
 
   const nullifiers = notes.map((n) => signer.createNullifier(n));
   const notesWithNullfiers = zip(notes, nullifiers).map(([n, nf]) =>